refactor: move Color and Theme to Theme package
